Create CONTRIBUTING.md
I am an example of a new contributor. I'm ready to start working on stuff, but I know nothing about the context here, and it's clear looking at it that there's a fair amount to understand.  As this project has been extremely quiet for quite some time I'm committing this into the repo in hopes that someone still attached to this project sees it and starts fleshing it out.